requirements.txt
ginger-dj==5.3.13 #ginger-society
mysqlclient==2.1.1
pyyaml==6.0.1
uritemplate==4.1.1
psycopg2==2.9.5
redis==4.5.5
prometheus-client==0.20.0
pytz==2024.1
inflection==0.5.1
packaging
setuptools
pydentic
certifi